% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{add_covariate}
\alias{add_covariate}
\title{add metadata to pb_cds from cds}
\usage{
add_covariate(ccs, pb_cds, covariate, colname = covariate)
}
\arguments{
\item{ccs}{a cell count set object}

\item{pb_cds}{a pseudobulked cds}

\item{covariate}{the column name in colData(ccs) to add to the pb_cds}
}
\description{
add metadata to pb_cds from cds
}
