% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregated_expr_data}
\alias{aggregated_expr_data}
\title{calculates the fraction expression, mean expression and specificity
for each gene split by the specified group}
\usage{
aggregated_expr_data(cds, group_cells_by = "cell_type_broad")
}
\arguments{
\item{group_cells_by}{}
}
\description{
calculates the fraction expression, mean expression and specificity
for each gene split by the specified group
}
