% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{assemble_transition_graph_from_perturbations}
\alias{assemble_transition_graph_from_perturbations}
\title{Assemble a state transition graph from a set of perturbations}
\usage{
assemble_transition_graph_from_perturbations(
  control_timeseries_ccm,
  perturbation_ccm_tbl,
  q_val = 0.01,
  start_time = NULL,
  stop_time = NULL,
  perturbation_col = "knockout",
  interval_col = "timepoint",
  interval_step = 2,
  min_interval = 4,
  max_interval = 24,
  log_abund_detection_thresh = -5,
  percent_max_threshold = 0,
  min_pathfinding_lfc = 1,
  links_between_components = c("none", "strongest-pcor", "strong-pcor"),
  verbose = FALSE,
  edge_whitelist = NULL,
  edge_blacklist = NULL,
  ...
)
}
\description{
This function takes as input a control timeseries Hooke model, and a set of timeseries perturbation models.
Each perturbation model describes a separate experimental perturbation that may eliminate one or more
cell states in the experiment. The tibble must have columns "perturb_name" and "perturb_ccm", and each row
must have a perturbation model with a unique name.
}
\details{
The function returns a state graph with edges annotated by the level of support from the perturbations.
}
