% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{bootstrap_ccs}
\alias{bootstrap_ccs}
\title{resamples the ccs counts using a multinomial distribution}
\usage{
bootstrap_ccs(ccs, random.seed = NULL)
}
\description{
resamples the ccs counts using a multinomial distribution
}
