% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{bootstrap_vhat}
\alias{bootstrap_vhat}
\title{computes the avg vhat across n bootstraps}
\usage{
bootstrap_vhat(
  ccs,
  full_model_formula_str,
  best_full_model,
  best_reduced_model,
  reduced_pln_model,
  pseudocount,
  initial_penalties,
  pln_min_ratio,
  pln_num_penalties,
  verbose,
  norm_method,
  num_bootstraps,
  backend
)
}
\description{
computes the avg vhat across n bootstraps
}
