% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{break_cycles_in_state_transition_graph}
\alias{break_cycles_in_state_transition_graph}
\title{Break cycles in a state graph}
\usage{
break_cycles_in_state_transition_graph(state_graph, support_attribute)
}
\description{
Break cycles in a state graph
}
