% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{build_interval_formula}
\alias{build_interval_formula}
\title{Builds a model formula for time series models based on the range of the data.
This is a utility function that puts the knots in reasonable positions based on the range of the data.}
\usage{
build_interval_formula(
  ccs,
  num_breaks,
  interval_var = "timepoint",
  interval_start = NULL,
  interval_stop = NULL
)
}
\arguments{
\item{character}{interval_var}

\item{numeric}{interval_stop Interval stop value.}
}
\value{
An interval model formula.
}
\description{
Builds a model formula for time series models based on the range of the data.
This is a utility function that puts the knots in reasonable positions based on the range of the data.
}
