% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{build_timeseries_transition_graph}
\alias{build_timeseries_transition_graph}
\title{Identify the possible origins for each destination}
\usage{
build_timeseries_transition_graph(
  ccm,
  extant_cell_type_df,
  pathfinding_graph,
  q_val = 0.01,
  start_time = NULL,
  stop_time = NULL,
  interval_col = "timepoint",
  interval_step = 2,
  min_interval = 4,
  max_interval = 24,
  min_pathfinding_lfc = 1,
  make_dag = FALSE,
  ...
)
}
\description{
Identify the possible origins for each destination
}
