% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{calc_shortest_path}
\alias{calc_shortest_path}
\title{Calculate the shortest path between two points}
\usage{
calc_shortest_path(G, from, to)
}
\arguments{
\item{edges}{data frame of edges with edge weights}
}
\value{
data frame containing the shortest path
}
\description{
Calculate the shortest path between two points
}
