% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_count_model.R
\name{cds}
\alias{cds}
\title{Get the underlying cell_data_set object from a cell_count_model.}
\usage{
cds(ccm)
}
\arguments{
\item{ccm}{A cell_count_model object.}
}
\value{
A cell_data_set object
}
\description{
Get the underlying cell_data_set object from a cell_count_model.
}
\examples{
\dontrun{
cds(ccm)
}
}
