% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\docType{class}
\name{cell_count_model}
\alias{cell_count_model}
\alias{cell_count_model-class}
\title{The cell_count_model class}
\description{
The main class used by Monocle3 to hold single-cell expression data.
cell_count_model extends the Bioconductor SingleCellExperiment class.
}
\details{
This class is initialized from a matrix of expression values along with cell
and feature metadata.
}
\section{Fields}{

\describe{
\item{\code{ccs}}{cell_count_set the underlying object of cell counts this class models.}

\item{\code{model_formula}}{specifies the model for the cell abundances.}

\item{\code{model_family}}{PLNnetworkfamily a family of PLNnetwork models for the cell count data.}

\item{\code{best_model}}{PLNnetworkfit the current best PLN network model for the cell count data.}

\item{\code{model_aux}}{SimpleList auxiliary information from from PLN model construction}
}}

