% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\docType{class}
\name{cell_count_set}
\alias{cell_count_set}
\alias{cell_count_set-class}
\title{The cell_count_set class}
\description{
The main class used by Hooke to hold cell abundances data.
cell_count_set extends the Monocle's cell_data_set class.
}
\details{
This class is initialized from a matrix of expression values along with cell
and feature metadata.
}
\section{Fields}{

\describe{
\item{\code{cds}}{cell_data_set, the Monocle cell data set object that this class models.}
}}

