% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_count_model.R
\name{centroids}
\alias{centroids}
\title{Get the centroids of cell groups in UMAP/PCA space.}
\usage{
centroids(ccs, reduction_method = "UMAP")
}
\arguments{
\item{ccs}{A cell_count_set object.}
}
\value{
A data frame of centroid coordinates
}
\description{
Get the centroids of cell groups in UMAP/PCA space.
}
