% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{classify_genes_over_graph}
\alias{classify_genes_over_graph}
\title{Classify each gene's pattern of expresison in each state in a state transition graph}
\usage{
classify_genes_over_graph(
  ccm,
  state_graph,
  gene_ids = NULL,
  group_nodes_by = NULL,
  log_fc_thresh = 1,
  abs_expr_thresh = 0.001,
  sig_thresh = 0.05,
  min_samples_detected = 2,
  min_cells_per_pseudobulk = 3,
  cores = 1,
  ...
)
}
\description{
Classify each gene's pattern of expresison in each state in a state transition graph
}
