% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{clean_pln_model_object}
\alias{clean_pln_model_object}
\title{function to get rid of extra stuff we don't ened to store}
\usage{
clean_pln_model_object(ccm)
}
\description{
function to get rid of extra stuff we don't ened to store
}
