% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{collect_pln_graph_edges}
\alias{collect_pln_graph_edges}
\title{Orient graph edges from a PLNnetwork using a contrast between conditions}
\usage{
collect_pln_graph_edges(
  ccm,
  cond_b_vs_a_tbl,
  log_abundance_thresh = 1 - 5,
  model_for_pcors = "reduced"
)
}
\arguments{
\item{ccm}{A cell_count_model}

\item{cond_b_vs_a_tbl}{A contrast between two conditions as returned by compare_abundances()}
}
\description{
Orient graph edges from a PLNnetwork using a contrast between conditions
}
