% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{compare_ko_to_wt_at_timepoint}
\alias{compare_ko_to_wt_at_timepoint}
\title{Helper function to plot kinetics}
\usage{
compare_ko_to_wt_at_timepoint(
  tp,
  perturbation_ccm,
  wt_pred_df,
  ko_pred_df,
  interval_col
)
}
\arguments{
\item{tp}{timepoint}

\item{perturbation_ccm}{a cell count model with a perturbation}

\item{wt_pred_df}{control output from estimate_abundances_over_interval()}

\item{ko_pred_df}{perturbation output from estimate_abundances_over_interval()}

\item{interval_col}{column that matches the timepoint information}
}
\description{
Helper function to plot kinetics
}
