% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contract_ccm}
\alias{contract_ccm}
\title{change cell count set grouping}
\usage{
contract_ccm(ccm, cell_group = NULL, sample_group = NULL)
}
\arguments{
\item{ccm}{a cell count model}

\item{cell_group}{string specifying how to aggregate the cell counts}
}
\description{
change cell count set grouping
}
