% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{contract_state_graph}
\alias{contract_state_graph}
\title{Simplify a directed state transition graph by grouping nodes according to a specified label}
\usage{
contract_state_graph(ccm, state_graph, group_nodes_by)
}
\description{
Simplify a directed state transition graph by grouping nodes according to a specified label
}
