% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{estimate_abundances}
\alias{estimate_abundances}
\title{Predict cell type abundances given a PLN model and a set of inputs for its covariates}
\usage{
estimate_abundances(ccm, newdata, min_log_abund = -5)
}
\arguments{
\item{ccm}{A cell_count_model.}

\item{newdata}{tibble A tibble of variables used for the prediction.}

\item{min_log_abund}{numeric Minimum log abundance value.}
}
\value{
A tibble of cell abundance predictions.
}
\description{
Predict cell type abundances given a PLN model and a set of inputs for its covariates
}
