% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{estimate_abundances_cond}
\alias{estimate_abundances_cond}
\title{Predict cell type abundances given a PLN model and a set of inputs for its covariates
and observed counts}
\usage{
estimate_abundances_cond(
  ccm,
  newdata,
  cond_responses,
  min_log_abund = -5,
  cell_group = "cell_group",
  type = c("link", "response"),
  pln_model = c("full", "reduced")
)
}
\arguments{
\item{ccm}{A cell_count_model.}

\item{newdata}{tibble A tibble of variables used for the prediction.
count set for ccm.}

\item{cond_responses}{a data frame containing the counts of the observed variables}

\item{min_log_abund}{numeric Minimum log abundance value.}

\item{cell_group}{string The name of the groups that are being estimated.}
}
\value{
A tibble of cell abundance predictions.
}
\description{
Predict cell type abundances given a PLN model and a set of inputs for its covariates
and observed counts
}
