% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{estimate_abundances_over_interval}
\alias{estimate_abundances_over_interval}
\title{Predict cell type abundances given a PLN model over a range of time or other interval}
\usage{
estimate_abundances_over_interval(
  ccm,
  interval_start,
  interval_stop,
  interval_col = "timepoint",
  interval_step = 2,
  min_log_abund = -5,
  newdata = tibble(),
  scale = c("log", "log10", "log2", "per_1000")
)
}
\arguments{
\item{ccm}{A cell_count_model.}

\item{interval_start}{numeric Interval start value.}

\item{interval_stop}{numeric Interval stop value.}

\item{interval_col}{character Interval values are taken from the interval_var data. Default is "timepoint".}

\item{interval_step}{numeric Interval size. Default is 2.}

\item{min_log_abund}{numeric Minimum log abundance value.}

\item{newdata}{tibble A tibble of variables used for the prediction.}
}
\value{
A tibble of cell abundance predictions.
}
\description{
Predict cell type abundances given a PLN model over a range of time or other interval
}
