% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_cds}
\alias{filter_cds}
\title{filters a cds}
\usage{
filter_cds(cds, ...)
}
\arguments{
\item{...}{expressions that return a logical value}
}
\value{
a cell data set object
}
\description{
filters a cds
}
