% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{find_deg_path}
\alias{find_deg_path}
\title{Find degs either along an entire path or between adjacent pairs}
\usage{
find_deg_path(
  ccm,
  cond_b_vs_a_tbl,
  p_value_threshold = 1,
  adj_pairs = FALSE,
  ...
)
}
\arguments{
\item{ccm}{a cell count model object}

\item{p_value_threshold}{p_value threshold for inclusion of edges}

\item{adj_pairs}{If TRUE, does the DEG testing between adjacent pairs. If FALSE, includes all states in the deg testing.}
}
\description{
Find degs either along an entire path or between adjacent pairs
}
