% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_children}
\alias{get_children}
\title{get the children of a state in a state transition graph}
\usage{
get_children(state_graph, cell_state)
}
\description{
get the children of a state in a state transition graph
}
