% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_colors}
\alias{get_colors}
\title{returns different color palettes}
\usage{
get_colors(num_colors, type = "rainbow")
}
\arguments{
\item{num_colors}{the number of colors needed}
}
\description{
returns different color palettes
}
