% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{get_extant_cell_types}
\alias{get_extant_cell_types}
\title{Return a dataframe that describes which cell types are present in a time interval}
\usage{
get_extant_cell_types(
  ccm,
  start,
  stop,
  interval_col = "timepoint",
  interval_step = 2,
  log_abund_detection_thresh = -5,
  pct_dynamic_range = 0.25,
  pct_range_detection_thresh = pct_dynamic_range,
  min_cell_range = 2,
  ...
)
}
\description{
Return a dataframe that describes which cell types are present in a time interval
}
