% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_count_model.R
\name{get_norm_counts}
\alias{get_norm_counts}
\title{return a matrix of normalized counts}
\usage{
get_norm_counts(ccs, round = FALSE)
}
\arguments{
\item{ccs}{cell_count_set}

\item{round}{TRUE if counts are rounded}
}
\description{
return a matrix of normalized counts
}
