% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_norm_df}
\alias{get_norm_df}
\title{function to easily plot to easily feed in normalized counts from ccs
into boxplots}
\usage{
get_norm_df(ccs)
}
\description{
function to easily plot to easily feed in normalized counts from ccs
into boxplots
}
