% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{get_paga_graph}
\alias{get_paga_graph}
\title{Extract a partitioned abstract graph from a Monocle cell_data_set object}
\usage{
get_paga_graph(cds, reduction_method = "UMAP", partition_q_value = 0.05)
}
\arguments{
\item{cds}{A cell_data_set object. cluster_cells() must have been called.}

\item{reduction_method}{The coordinate space in which to build the graph}
}
\description{
QUESTION: Do the cells need to be grouped by monocle cluster? Or can they
be grouped arbitrarily (e.g. by cell type? Would be good to generalize if
possible. For now, only works when cells are grouped by Monocle cluster
}
