% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_parents}
\alias{get_parents}
\title{get the parent(s) of a state in a state transition graph}
\usage{
get_parents(state_graph, cell_state)
}
\description{
get the parent(s) of a state in a state transition graph
}
