% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_shortest_path}
\alias{get_shortest_path}
\title{calculates the shortest weighted path between two points}
\usage{
get_shortest_path(from, to, traversal_graph)
}
\arguments{
\item{target}{}
}
\description{
calculates the shortest weighted path between two points
}
