% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{get_siblings}
\alias{get_siblings}
\title{get the siblings of a state in a state transition graph}
\usage{
get_siblings(state_graph, cell_state)
}
\description{
get the siblings of a state in a state transition graph
}
