% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_weighted_edges}
\alias{get_weighted_edges}
\title{weights edges based on partial correlation,
change in abundance, and transcriptomic distance}
\usage{
get_weighted_edges(ccm, edges, alpha = 1, beta = 1, gamma = 1, sum_weights = T)
}
\description{
weights edges based on partial correlation,
change in abundance, and transcriptomic distance
}
