% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{init_penalty_matrix}
\alias{init_penalty_matrix}
\title{Initialize the PLN network penalty matrix, accepting optional whitelists and
blacklists of edges that are "free" or "off limits" between cell groups}
\usage{
init_penalty_matrix(
  ccs,
  whitelist = NULL,
  blacklist = NULL,
  base_penalty = 1,
  min_penalty = 0.01,
  max_penalty = 1e+06
)
}
\arguments{
\item{ccs}{A cell_count_set of aggregated cell counts}

\item{whitelist}{a data frame with two columns corresponding to (undirected) edges that should receive no penalty}

\item{blacklist}{a data frame with two columns corresponding to (undirected) edges that should receive very high penalty}

\item{dist_fun}{A function that returns a penalty based given a distance between two clusters}
}
\description{
Initialize the PLN network penalty matrix, accepting optional whitelists and
blacklists of edges that are "free" or "off limits" between cell groups
}
