% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{initial_pcor_graph}
\alias{initial_pcor_graph}
\title{Get an initial graph for use as a whitelist in fitting a cell count model}
\usage{
initial_pcor_graph(ccs)
}
\arguments{
\item{ccs}{A cell_count_model object.}
}
\description{
Get an initial graph for use as a whitelist in fitting a cell count model
}
