% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{measure_perturbation_freq_along_path}
\alias{measure_perturbation_freq_along_path}
\title{score a path based on fitting a linear model of perturbation ~ geodesic distance}
\usage{
measure_perturbation_freq_along_path(
  path_df,
  ccs,
  cells_along_path_df,
  perturbation_col = "knockout",
  interval_col = "timepoint"
)
}
\description{
score a path based on fitting a linear model of perturbation ~ geodesic distance
}
