% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_assembly.R
\name{measure_time_delta_along_path}
\alias{measure_time_delta_along_path}
\title{score a path based on fitting a linear model of time ~ geodeisic distance}
\usage{
measure_time_delta_along_path(
  path_df,
  ccs,
  cells_along_path_df,
  interval_col = "timepoint"
)
}
\description{
score a path based on fitting a linear model of time ~ geodeisic distance
}
