% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-cell_count_model.R
\name{model}
\alias{model}
\title{Get the selected PLNnetwork from a cell_count_model object.}
\usage{
model(ccm, model_to_return = c("full", "reduced"))
}
\arguments{
\item{ccm}{A cell_count_model object.}
}
\value{
An updated cell_count_model object
}
\description{
Get the selected PLNnetwork from a cell_count_model object.
}
\examples{
\dontrun{
model(ccm)
}
}
