% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{my_plot_cells}
\alias{my_plot_cells}
\title{rename this later?
probably needs fixing}
\usage{
my_plot_cells(
  data,
  color_cells_by = "cluster",
  cell_size = 1,
  legend_position = "none",
  residuals = NULL,
  cond_b_vs_a_tbl = NULL
)
}
\description{
rename this later?
probably needs fixing
}
