% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{new_cell_count_set}
\alias{new_cell_count_set}
\title{Create a new cell_data_set object.}
\usage{
new_cell_count_set(
  cds,
  sample_group,
  cell_group,
  sample_metadata = NULL,
  cell_metadata = NULL
)
}
\arguments{
\item{cds}{A Monocle cell data set object.}

\item{sample_group}{A column in colData(cds) that specifes how cells are grouped into samples}

\item{cell_group}{A column in colData(cds) that specifies how cells are grouped into types or states (e.g. cluster)}

\item{sample_metadata}{data frame containing attributes of individual samples, where}

\item{cell_metadata}{data frame containing attributes of individual cell groups, where
\code{row.names(cell_metadata)} are entries in \code{cell_group}}
}
\value{
a new cell_data_set object
}
\description{
Create a new cell_data_set object.
}
