% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot a UMAP colored by how cells shift in a given contrast}
\usage{
plot_abundance(
  ccs,
  cond_b_vs_a_tbl,
  mask = list(),
  cell_size = 1,
  q_value_thresh = 1,
  fc_limits = c(-3, 3),
  plot_labels = c("significant", "all", "none"),
  alpha = 1,
  group_label_size = 2,
  x = 1,
  y = 2,
  downsample = NULL
)
}
\arguments{
\item{ccs}{A cell_count_set object.}

\item{cond_b_vs_a_tbl}{data.frame A data frame from compare_abundances.}

\item{mask}{a list of cell types to gray out in the plots}

\item{cell_size}{numeric The size of cells in the plot.}

\item{q_value_thresh}{numeric Remove contrasts whose change in
q-value exceeds q_value_thresh.}

\item{fc_limits}{vector The range of cell abundance changes to
include in the plot.}

\item{plot_labels}{string Choose cell groups to label.}

\item{group_label_size}{numeric The size of group labels in the plot.}

\item{x}{numeric The column number for the UMAP x coordinate.}

\item{y}{numeric The column number for the UMAP y coordinate.}

\item{downsample}{how much to downsample the plots}
}
\description{
Plot a UMAP colored by how cells shift in a given contrast
}
