% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cells_per_sample}
\alias{plot_cells_per_sample}
\title{helper function for adding attributes to network node
assumes gene id is in the dataframe}
\usage{
plot_cells_per_sample(
  ccs,
  x_col,
  normalize = F,
  plot_zeroes = F,
  plot_points = F,
  log_scale = F,
  nrow = 1,
  legend_position = "none"
)
}
\description{
helper function for adding attributes to network node
assumes gene id is in the dataframe
}
