% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_contrast}
\alias{plot_contrast}
\title{Plot a UMAP colored by how cells shift in a given contrast}
\usage{
plot_contrast(
  ccm,
  cond_b_vs_a_tbl,
  log_abundance_thresh = -5,
  scale_shifts_by = c("receiver", "sender", "none"),
  edge_size = 2,
  cell_size = 1,
  q_value_thresh = 1,
  group_label_size = 2,
  plot_labels = c("significant", "all", "none"),
  fc_limits = c(-3, 3),
  sender_cell_groups = NULL,
  receiver_cell_groups = NULL,
  plot_edges = c("all", "directed", "undirected", "none"),
  label_cell_groups = list(),
  repel_labels = TRUE,
  model_for_pcors = "reduced",
  switch_label = NULL,
  sub_cds = NULL,
  alpha = 1,
  x = 1,
  y = 2
)
}
\arguments{
\item{ccm}{A cell_count_model object.}

\item{cond_b_vs_a_tbl}{data.frame A data frame from compare_abundances.}
}
\value{
A ggplot2 plot object.
}
\description{
Plot a UMAP colored by how cells shift in a given contrast
}
