% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_map}
\alias{plot_map}
\title{plots the resulting path as an igraph
instead of as a UMAP}
\usage{
plot_map(data, edges, color_nodes_by = "", arrow.gap = 0.02, scale = F)
}
\description{
plots the resulting path as an igraph
instead of as a UMAP
}
