% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_origins}
\alias{plot_origins}
\title{Simple function for inspecting origins for each cell state}
\usage{
plot_origins(
  data,
  path_df = path_df,
  edge_size = 2,
  path_color = "black",
  color_cells_by = "cluster",
  color_path_by = NULL,
  cond_b_vs_a_tbl = NULL
)
}
\description{
Simple function for inspecting origins for each cell state
}
