% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_path}
\alias{plot_path}
\title{plots a path on top of}
\usage{
plot_path(
  data,
  state_graph,
  edge_size = 1,
  path_color = "black",
  color_path_by = NULL,
  cond_b_vs_a_tbl = NULL,
  directed = TRUE,
  x = 1,
  y = 2,
  switch_label = NULL,
  ...
)
}
\description{
plots a path on top of
}
