% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_single_gene}
\alias{plot_single_gene}
\title{code to plot a single gene nicely and save w a transparent background}
\usage{
plot_single_gene(cds, gene, save_file = F, file_path = NULL, x = 1, y = 2)
}
\arguments{
\item{file_path}{}
}
\description{
code to plot a single gene nicely and save w a transparent background
}
