% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_graph_gene_expression}
\alias{plot_state_graph_gene_expression}
\title{Plot the expression of a single gene on the state transition graph}
\usage{
plot_state_graph_gene_expression(
  ccm,
  state_graph,
  genes,
  method = "min",
  gene_expr = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  edge_labels = NULL,
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  node_size = 6,
  num_layers = 10,
  edge_size = 0.5,
  fract_expr = 0,
  mean_expr = 0,
  unlabeled_groups = c("Unknown"),
  scale_to_range = FALSE,
  hide_unlinked_nodes = TRUE,
  switch_label = NULL,
  label_font_size = 6,
  label_conn_linetype = "dotted",
  legend_position = "none",
  group_outline = FALSE
)
}
\description{
Plot the expression of a single gene on the state transition graph
}
