% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_graph_losses}
\alias{plot_state_graph_losses}
\title{This function plots when each cell state is lost following a perturbation}
\usage{
plot_state_graph_losses(
  perturbation_ccm,
  state_graph,
  start_time,
  stop_time,
  interval_step,
  interval_col,
  log_abund_detection_thresh,
  q_val,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  label_edges_by = NULL,
  edge_weights = NULL,
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  node_size = 2,
  num_layers = 10,
  edge_size = 2,
  fract_expr = 0,
  mean_expr = 0,
  unlabeled_groups = c("Unknown"),
  hide_unlinked_nodes = TRUE,
  label_font_size = 6,
  label_conn_linetype = "dotted",
  legend_position = "none",
  group_outline = FALSE,
  ...
)
}
\description{
FIXME: this is really a prototype and needs lots of cleanup and refactoring.
It could plot the earliest time a loss is detected, the latest, etc. But
right now it only plots the loss at the time of peak abundance in the control
see estimate_loss_timing() for more details.
}
