% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_transition_graph}
\alias{plot_state_transition_graph}
\title{Plot a graph that summarized an assembled state transition graph}
\usage{
plot_state_transition_graph(
  ccm,
  edges,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  layer_nodes_by = NULL,
  cond_b_v_a_tbl = NULL,
  fc_limits = NULL,
  genotype_models_tbl = NULL,
  concordant = TRUE,
  genes = list(),
  labels = NULL,
  qval_threshold = 0.05,
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  node_size = 2,
  num_layers = 10,
  edge_size = 0.5,
  fract_expr = 0,
  mean_expr = 0,
  method = "min",
  unlabeled_groups = c("Unknown"),
  hide_unlinked_nodes = TRUE,
  label_font_size = 6,
  label_conn_linetype = "dotted",
  legend_position = "none",
  group_outline = FALSE
)
}
\description{
Plot a graph that summarized an assembled state transition graph
}
