% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_sub_contrast}
\alias{plot_sub_contrast}
\title{wrapper for plot contrast that allows you to facet the plot
you can also switch between sub + full umap space}
\usage{
plot_sub_contrast(
  ccm,
  cond_b_vs_a_tbl,
  facet_group = "major_group",
  select_group = NULL,
  log_abundance_thresh = -5,
  edge_size = 2,
  cell_size = 1,
  q_value_thresh = 1,
  group_label_size = 2,
  plot_labels = c("significant", "all", "none"),
  plot_edges = c("all", "directed", "undirected", "none"),
  fc_limits = c(-3, 3),
  sub_space = T,
  ...
)
}
\arguments{
\item{facet_group}{}
}
\description{
wrapper for plot contrast that allows you to facet the plot
you can also switch between sub + full umap space
}
