% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{pseudobulk_ccs_for_states}
\alias{pseudobulk_ccs_for_states}
\title{Compute a pseudobulk expression matrix for a ccs}
\usage{
pseudobulk_ccs_for_states(
  ccs,
  state_col = NULL,
  collapse_samples = FALSE,
  gene_group_df = NULL,
  agg_rowdata = NULL,
  cell_agg_fun = "mean",
  gene_agg_fun = "sum",
  norm_method = "size_only",
  scale_agg_values = FALSE,
  pseudocount = 0
)
}
\arguments{
\item{ccs}{a cell count set object}

\item{state_col}{column to aggregate expression. Defaults using the cell_group used in ccs construction.}

\item{collapse_samples}{boolean Whether to collapse sample groups into one.}

\item{gene_group_df}{A dataframe in which the first column contains gene ids or short gene names and the second contains group membership for creating metagene. If NULL, genes are not grouped.}

\item{agg_rowdata}{A dataframe with metagene information, with the rows equal to the number of gene groupings. Must be specified if \code{gene_group_df} is not NULL. Otherwise this is ignored.}
}
\description{
Compute a pseudobulk expression matrix for a ccs
}
