% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{pseudobulk_cds_for_states}
\alias{pseudobulk_cds_for_states}
\title{Compute a pseudobulk expression matrix for a model}
\usage{
pseudobulk_cds_for_states(ccm, state_col = NULL, collapse_samples = FALSE)
}
\description{
Compute a pseudobulk expression matrix for a model
}
