% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_projection}
\alias{run_projection}
\title{projection wrap up}
\usage{
run_projection(query_cds, ref_cds, directory_path, transfer_type = "cluster")
}
\arguments{
\item{directory_path}{a string giving the name of the directory from which to read the model files}
}
\description{
projection wrap up
}
