% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{select_model}
\alias{select_model}
\title{Select the model a cell_count_model should use}
\usage{
select_model(ccm, criterion = "EBIC", sparsity_factor = 1)
}
\arguments{
\item{ccm}{A cell_count_model object.}

\item{criterion}{a character string specifying the PLNmodels criterion to use. Must be one of "BIC", "EBIC" or "StARS".}
}
\value{
an updated cell_count_model object
}
\description{
Select the model a cell_count_model should use
}
