% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_umap_space}
\alias{switch_umap_space}
\title{this switches the umap space to the sub umap space}
\usage{
switch_umap_space(cds, sub_space = TRUE)
}
\description{
this switches the umap space to the sub umap space
}
