% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_gene_pattern}
\alias{top_gene_pattern}
\title{score DEGs on specificity}
\usage{
top_gene_pattern(
  ccm,
  gene_patterns_over_cell_graph,
  pattern = NULL,
  grep_pattern = NULL,
  cell_group = "cell_type_sub",
  n = 5,
  return_genes = T
)
}
\arguments{
\item{cell_group}{}
}
\description{
score DEGs on specificity
}
