% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{compare_abundances}
\alias{compare_abundances}
\title{Compare two estimates of cell abundances from a Hooke model.}
\usage{
compare_abundances(
  ccm,
  cond_x,
  cond_y,
  by = "cell_group",
  method = c("BH", "bonferroni", "hochberg", "hommel", "BY")
)
}
\arguments{
\item{ccm}{A cell_count_model.}

\item{cond_x}{tibble A cell type abundance estimate from estimate_abundances().}

\item{cond_y}{tibble A cell type abundance estimate from estimate_abundances().}

\item{by}{string The column name used to join the two estimates.}

\item{method}{string A method for correcting P-value multiple comparisons.
This can be "BH" (Benjamini & Hochberg), "bonferroni" (Bonferroni),
"hochberg" (Hochberg), "hommel", (Hommel), or "BYH" (Benjamini & Yekutieli).}
}
\value{
tibble A table contrasting cond_x and cond_y (interpret as Y/X).
}
\description{
Compare two estimates of cell abundances from a Hooke model.
}
