% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{compare_abundances}
\alias{compare_abundances}
\title{Compare two estimates of cell abundances from a Hooke model}
\usage{
compare_abundances(ccm, cond_x, cond_y, method = "BH")
}
\arguments{
\item{ccm}{A cell_count_model}

\item{cond_x}{An estimate from estimate_abundances()}

\item{cond_y}{An estimate from estimate_abundances()}
}
\value{
A table contrasting cond_x and cond_y (interpret as Y/X)
}
\description{
Compare two estimates of cell abundances from a Hooke model
}
