% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{estimate_abundances_over_interval}
\alias{estimate_abundances_over_interval}
\title{Predict cell type abundances given a PLN model over a range of time or other interval}
\usage{
estimate_abundances_over_interval(
  ccm,
  start,
  stop,
  interval_col = "timepoint",
  interval_step = 2,
  ...
)
}
\description{
Predict cell type abundances given a PLN model over a range of time or other interval
}
