% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{new_cell_count_model}
\alias{new_cell_count_model}
\title{Create a new cell_count_model object.}
\usage{
new_cell_count_model(
  ccs,
  model_formula_str,
  penalty_matrix = NULL,
  whitelist = NULL,
  blacklist = NULL,
  sparsity_factor = 0.1,
  base_penalty = 1,
  min_penalty = 0.01,
  max_penalty = 1e+06,
  verbose = FALSE,
  pseudocount = 0,
  pln_min_ratio = 0.001,
  pln_num_penalties = 30,
  ...
)
}
\arguments{
\item{ccs}{A Hooke cell_count_set object.}

\item{model_formula_str}{A character string specifying the model of cell abundances across samples.
where terms refer to columns in\code{colData(ccs)}}

\item{penalty_matrix}{A numeric NxN symmetric matrix specifying penalties for
the PLN model, where N is the number of cell types. Entries must be
positive. Use to specify an undirected graph prior for the PLN model.}

\item{sparsity_factor}{A positive number to control how sparse the PLN network is. Larger values make the network more sparse.}
}
\value{
a new cell_count_model object
}
\description{
Fits a PLNnetwork according to a formula. Accepts a matrix of penalties as a
way of encoding a graph prior. Automatically selects sparsity parameter, but
allows user to update it.
}
