% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_model.R
\name{new_cell_count_set}
\alias{new_cell_count_set}
\title{Create a new cell_data_set object.}
\usage{
new_cell_count_set(
  cds,
  sample_group,
  cell_group,
  sample_metadata = NULL,
  cell_metadata = NULL,
  lower_threshold = NULL,
  upper_threshold = NULL
)
}
\arguments{
\item{cds}{A Monocle cell data set object.}

\item{sample_group}{A column in colData(cds) that specifes how cells are grouped into samples.}

\item{cell_group}{A column in colData(cds) that specifies how cells are grouped into types or states (e.g. cluster).}

\item{sample_metadata}{Data frame containing attributes of individual samples, where
the column named 'sample' has entries in \code{sample_group}.}

\item{cell_metadata}{Data frame containing attributes of individual cell groups, where
\code{row.names(cell_metadata)} are entries in \code{cell_group}}

\item{lower_threshold}{numeric Minimum number of cells in retained cell_groups.}

\item{upper_threshold}{numeric Maximum number of cells in retained cell_groups.}
}
\value{
a new cell_data_set object
}
\description{
Create a new cell_data_set object.
}
