% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cells_per_sample}
\alias{plot_cells_per_sample}
\title{Plot Cells Per Sample}
\usage{
plot_cells_per_sample(
  ccs,
  x_col,
  y_col = c("count", "count_per_1000"),
  cell_groups = c(),
  batch_col = NULL,
  color_by = "cell_group",
  plot_zeroes = T,
  plot_points = F,
  facet = T,
  log_scale = F,
  nrow = 1,
  legend_position = "none"
)
}
\arguments{
\item{ccs}{A data frame or tibble containing cell count data.}

\item{x_col}{A string specifying the column name to be used for the x-axis.}

\item{y_col}{A string specifying the column name to be used for the y-axis. Default is "count".}

\item{cell_groups}{A vector of cell group names to filter the data. Default is an empty vector.}

\item{color_by}{A string specifying the column name to be used for coloring the plot. Default is "cell_group".}

\item{plot_zeroes}{A logical value indicating whether to plot zero counts. Default is FALSE.}

\item{plot_points}{A logical value indicating whether to plot individual points. Default is FALSE.}

\item{log_scale}{A logical value indicating whether to use a logarithmic scale for the y-axis. Default is FALSE.}

\item{nrow}{An integer specifying the number of rows for faceting. Default is 1.}

\item{legend_position}{A string specifying the position of the legend. Default is "none".}
}
\value{
A ggplot object representing the boxplot of cell counts per sample.
}
\description{
This function generates a boxplot of cell counts per sample, with options to customize the plot.
}
