% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_contrast}
\alias{plot_contrast}
\title{Plot a UMAP colored by how cells shift in a given contrast}
\usage{
plot_contrast(
  ccm,
  cond_b_vs_a_tbl,
  log_abundance_thresh = -5,
  scale_shifts_by = c("receiver", "sender", "none"),
  edge_size = 2,
  cell_size = 1
)
}
\arguments{
\item{ccm}{A cell_count_model object.}

\item{criterion}{a character string specifying the PLNmodels criterion to use. Must be one of "BIC", "EBIC" or "StARS".}
}
\value{
an updated cell_count_model object
}
\description{
Plot a UMAP colored by how cells shift in a given contrast
}
