% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_graph_abundance_changes}
\alias{plot_state_graph_abundance_changes}
\title{Plot the expression of a single gene on the state transition graph}
\usage{
plot_state_graph_abundance_changes(
  ccm,
  state_graph,
  comp_abund_table,
  contrast = "contrast",
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  label_edges_by = NULL,
  edge_weights = NULL,
  fc_limits = c(-3, 3),
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  node_size = 6,
  min_edge_size = 0.1,
  max_edge_size = 2,
  unlabeled_groups = c("Unknown"),
  label_subset = NULL,
  label_groups = TRUE,
  hide_unlinked_nodes = TRUE,
  group_label_font_size = 6,
  edge_label_font_size = 2,
  label_conn_linetype = "dotted",
  legend_position = "none",
  con_colour = "darkgrey",
  group_outline = FALSE,
  flip_x = FALSE
)
}
\description{
Plot the expression of a single gene on the state transition graph
}
