% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_graph_gene_expression}
\alias{plot_state_graph_gene_expression}
\title{Plot the expression of a single gene on the state transition graph}
\usage{
plot_state_graph_gene_expression(
  ccs,
  state_graph,
  genes,
  method = "min",
  gene_expr = NULL,
  fract_expr = 0,
  mean_expr = 0,
  scale_to_range = FALSE,
  color_nodes_by = NULL,
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  label_edges_by = NULL,
  edge_weights = NULL,
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  min_node_size = 0.25,
  max_node_size = 2,
  min_edge_size = 0.1,
  max_edge_size = 2,
  unlabeled_groups = c("Unknown"),
  label_groups = TRUE,
  hide_unlinked_nodes = TRUE,
  group_label_font_size = 6,
  edge_label_font_size = 2,
  label_conn_linetype = "dotted",
  legend_position = "none",
  con_colour = "darkgrey",
  group_outline = FALSE
)
}
\description{
Plot the expression of a single gene on the state transition graph
}
