% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_state_graph_losses}
\alias{plot_state_graph_losses}
\title{This function plots when each cell state is lost following a perturbation}
\usage{
plot_state_graph_losses(
  perturbation_ccm,
  state_graph,
  start_time,
  stop_time,
  interval_step,
  interval_col,
  log_abund_detection_thresh,
  q_val,
  loss_time = c("largest_loss", "largest_loss_time", "earliest_time", "latest_time",
    "peak_loss_time"),
  label_nodes_by = NULL,
  group_nodes_by = NULL,
  label_edges_by = NULL,
  edge_weights = NULL,
  arrow.gap = 0.03,
  arrow_unit = 2,
  bar_unit = 0.075,
  node_size = 2,
  min_edge_size = 0.1,
  max_edge_size = 2,
  unlabeled_groups = c("Unknown"),
  label_subset = NULL,
  label_groups = TRUE,
  hide_unlinked_nodes = TRUE,
  group_label_font_size = 6,
  edge_label_font_size = 2,
  label_conn_linetype = "dotted",
  legend_position = "none",
  con_colour = "darkgrey",
  group_outline = FALSE,
  ...
)
}
\description{
FIXME: this is really a prototype and needs lots of cleanup and refactoring.
It could plot the earliest time a loss is detected, the latest, etc. But
right now it only plots the loss at the time of peak abundance in the control
see estimate_loss_timing() for more details.
}
